+++ demo/main.go
@@ -6,10 +6,11 @@ import (
 
 	"github.com/babylonchain/babylon-finality-gadget/btcclient"
 	"github.com/babylonchain/babylon-finality-gadget/sdk"
+	"github.com/babylonchain/babylon-finality-gadget/sdk/config"
 )
 
 func checkBlockFinalized(height uint64, hash string) {
-	client, err := sdk.NewClient(&sdk.Config{
+	client, err := sdk.NewClient(&config.Config{
 		ChainType: 0,
 		// TODO: avoid using stub contract
 		ContractAddr: "bbn1ghd753shjuwexxywmgs4xz7x2q732vcnkm6h2pyv9s6ah3hylvrqxxvh0f",
