+++ sdk/sdk_test.go
@@ -4,6 +4,7 @@ import (
 	"sort"
 	"testing"
 
+	"github.com/babylonchain/babylon-finality-gadget/sdk/config"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,7 +16,7 @@ import (
  */
 
 func newE2eClientWithStubContract() *BabylonFinalityGadgetClient {
-	stubContractConfig := &Config{
+	stubContractConfig := &config.Config{
 		ChainType:    0,
 		ContractAddr: "bbn1ghd753shjuwexxywmgs4xz7x2q732vcnkm6h2pyv9s6ah3hylvrqxxvh0f",
 	}
